cleaning up lib.rs
